Add href on git: and svn: URIs
The URLs used by GIT and SVN are perfectly valid URI, while not normalized there is no reason not to link them correctly to allow people to click them and have friendly OSes handle them correctly.
Haiku at least will do with the checkitout tool that opens a file requester and runs the scm checkout command in a Terminal.

Thanks to :%s/git:[a-zA-Z./~-]*/<a href="&">&<\/a>/gc
this patch fixes http://ffmpeg.org/download.html