web: reorder download page
Per my mail on the devel list from 2013-10-30, this change moves
the static builds further up in the download page and cleans up
the lists, in order to make it easier for users to discover the
downloads. Almost like a "Download here" button many users are
looking for.

Many casual users will be scared by the thought of having to compile
software, and when we send them to the download page, it's often
not apparent that the static builds work just as well.

The packages for Linux distributions have been moved down since
they contain older versions. Especially for Ubuntu, I added a
note about it being stuck at 0.10, which new users often don't
realize.

I also added links to the OS X package managers that enable users
to install ffmpeg directly.

For power users, the source code access is still there. Someone
willing to get the source will immediately see the git command.
I also changed the "people are not afraid to use them" to a more
direct statement.

Finally, a link to the wiki has been added, to point people to our
compilation guides, which are well maintained and a good reference.