Revert "web: move RSS generation to a seperate shell script"
This reverts commit f4c4145a82d69a24292df4c02b8b8ebe59c4884f.

Execution of scripts from the git checkout during make is not
supported currently. The Makefile is subject to security checks
on the server, these do not extend to external scripts.
It seems none of the root admins had time until now to extend
these security checks to external scripts and noone else volunteered
to fix them either (which isnt that trivial as the checks are shared
between several projects)

Thus to restore the automated website build on push, this commit
is reverted
Better solutions welcome, patches welcome, but dont break the
website build please, also if you do break it, revert, dont
just leave it assuming someone else will fix it.