web/donations: Remove ffis
There are some problems with the ffis webpage "Unable to connect to SQL server"
I tried to report this but it seems my email was not achieving that.
It may be safer to remove the link to ffis until we understand what exactly
is going on on the ffis side.

Suggested-by: Kieran
Reviewed-by: Thilo Borgmann <thilo.borgmann@mail.de>
Signed-off-by: Michael Niedermayer <michael@niedermayer.cc>