web/download: link to the latest release instead of a snapshot in the main download button
As suggested on IRC. There's already an Snapshot section right below it, and
the links to the releases are much lower and harder to find.

Also, remove the massive signing key button and move it to the snapshot
section. There's a "Release Verification" section below with a detailed
explanation and yet another link to the key.

Signed-off-by: James Almer <jamrial@gmail.com>